updated uiux interface
