+++ data/inventory.csv
@@ -1,0 +1,3 @@
+sku,name,quantity,supplier_id
+ABC-1001,Apples,10,SUP1
+XYZ-2002,Bananas,3,SUP2
