--- simple-inventory-system/Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-pandas = "*"
-customtkinter = "*"
-
-[dev-packages]
-black = "*"
-pytest = "*"
-
-[requires]
-python_version = "3.13"
-python_full_version = "3.13.7"
